linux-gen: ipc: remove unnecessary odp_packet_io_ipc_internal.h header
Copy the remaining necessary code from odp_packet_io_ipc_internal.h to
ipc.c.
