--- platform/linux-generic/include/odp_packet_io_ipc_internal.h
@@ -1,46 +1,0 @@
-/* Copyright (c) 2015-2018, Linaro Limited
- * All rights reserved.
- *
- * SPDX-License-Identifier:     BSD-3-Clause
- */
-
-#include <odp/api/packet_io.h>
-#include <odp_packet_io_internal.h>
-#include <odp/api/packet.h>
-#include <odp_packet_internal.h>
-#include <odp/api/shared_memory.h>
-
-#include <string.h>
-#include <unistd.h>
-#include <stdlib.h>
-
-/* number of odp buffers in odp ring queue */
-#define PKTIO_IPC_ENTRIES 4096
-
-#define PKTIO_IPC_ENTRY_MASK (PKTIO_IPC_ENTRIES - 1)
-
-/* that struct is exported to shared memory, so that processes can find
- * each other.
- */
-struct pktio_info {
-	struct {
-		/* number of buffer*/
-		int num;
-		/* size of packet/segment in remote pool */
-		uint32_t block_size;
-		char pool_name[ODP_POOL_NAME_LEN];
-		/* 1 if master finished creation of all shared objects */
-		int init_done;
-	} master;
-	struct {
-		void *base_addr;
-		uint32_t block_size;
-		char pool_name[ODP_POOL_NAME_LEN];
-		/* pid of the slave process written to shm and
-		 * used by master to look up memory created by
-		 * slave
-		 */
-		int pid;
-		int init_done;
-	} slave;
-} ODP_PACKED;
